+++ FRAuth/FRAuth/Callbacks/DeviceProfileCallback.swift
@@ -62,7 +62,7 @@ import Foundation
     func prepareCollectors() {
         if locationRequired {
             for deviceCollector in FRDeviceCollector.shared.collectors {
-                if String(describing: deviceCollector) == "FRProximity.LocationCollector" {
+                if String(describing: deviceCollector).contains("FRProximity.LocationCollector") {
                     self.locationCollector = deviceCollector
                 }
             }
@@ -77,7 +77,7 @@ import Foundation
         
         if metadataRequired {
             for deviceCollector in FRDeviceCollector.shared.collectors {
-                if String(describing: deviceCollector) == "FRProximity.BluetoothCollector" {
+                if String(describing: deviceCollector).contains("FRProximity.BluetoothCollector") {
                     self.profileCollector.collectors.append(deviceCollector)
                 }
             }