+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 - `Node` now supports `pageHeader`, and `pageDescription` attributes from `Page Node`. [SDKS-517]
 - `NumberAttributeInputCallback`, and `BooleanAttributeInputCallback` are now supported for IDM integration `Callback`. [SDKS-494]
 - `AbstractValidatedCallback` supports updated `Policies` structures. [SDKS-460]
+- `FRProximity.setLocationAccuracy` is added to specify `CLLocationManager.desiredAccuracy` configuration used in `LocationCollector`. [SDKS-617]
 
 #### Changed
 - `FRUI` now longer asks for user's consent when `DeviceProfileCallback` is the only `Callback` in the `Node`. [SDKS-436]
 - `FRAuth` was mistakenly allowing other app's private Keychain Access storage when .entitlement is misconfigured. [SDKS-552]
+- `FRProximity` SDK's `LocationCollector` now requests for Location Authorization while collecting Device Profile information if the authorization has not been asked yet. [SDKS-617]
 
 #### Deprecated